Update ML_2012_BE_eusilc_cs.do
Removed unemployed and self-employed.
MISSOC says that "special regulations for unemployed workers and for disabled for benefits for unemployed" but does not provide any amounts. I was also unable to find the amounts in LP&R. 
No info for the amounts for self-employed benefits were given for either data source.